j'ai ajoute d'autres fonctions et aussi quelque modifications

+++ E_Bank_Solution/src/main/java/com/example/E_Bank_Solution/Model/User.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userId;
+    private long userId;
     private String nom;
     private Integer age;
